requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
seaborn
opencv-python
scikit-learn
pillow
jupyter
